Handle Out-of-Bounds Features More Robustly
- Renamed `treat_data` to `treat_real_data` for improved clarity. This
function now raises an error if it encounters an out-of-bounds feature,
as the sampling pipeline should never generate a sample outside the
defined bounds of the design space.
- Introduced a new function, `select_in_bounds_feat`, to manage
out-of-bounds features. This function is intended for use when importing
new data, such as from a CSV file at the begining of a pipeline.
- Updated `DataTreatment` methods to explicitly require the `df_real`
argument when they are designed to process only real data and not scaled
data.